fix: load Google Analytics so GA4 receives data
The gtag.js scripts were commented out in _document.js, so window.gtag never
existed and every initGA/pageview/event call silently no-opped — GA4 reported
that it had never received any data from the site.

Uncommenting alone is not enough: next/script with afterInteractive does not
reliably inject from _document. Move the GA scripts to _app.js (the documented
placement), which makes window.gtag available. Verified in a real browser that
gtag loads and that accepting the cookie banner pushes an analytics_storage
'granted' consent update, after which hits flow.
